% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.bertini.R
\name{is.bertini}
\alias{is.bertini}
\title{Bertini Object Check}
\usage{
is.bertini(x)
}
\arguments{
\item{x}{object to be tested}
}
\value{
Vector of logicals.
}
\description{
Test whether an object is an bertini object.
}
\examples{

soln <- poly_solve("x^2 - 1 == 0")
is.bertini(soln)


}
