% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.bertini_input.R
\name{is.bertini_input}
\alias{is.bertini_input}
\title{Bertini_Input Object Check}
\usage{
is.bertini_input(x)
}
\arguments{
\item{x}{object to be tested}
}
\value{
Vector of logicals.
}
\description{
Test whether an object is an bertini_input object.
}
\examples{

if (has_bertini()) {

polys <- mp(c("x^2 + y^2 - 1", "x - y"))
struct <- bertini_input(polys)
is.bertini_input(struct)

}

}
