% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.bertini_posdim.R
\name{is.bertini_posdim}
\alias{is.bertini_posdim}
\title{Bertini_Posdim Object Check}
\usage{
is.bertini_posdim(x)
}
\arguments{
\item{x}{object to be tested}
}
\value{
Vector of logicals.
}
\description{
Test whether an object is an bertini_posdim object.
}
\examples{

if (has_bertini()) {
input <- bertini_input("x*y")
input <- modify_config(input, tracktype = 1)

out <- bertini(input, output = "pos_dim")
is.bertini_posdim(out)

}

}
