% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.witness_set.R
\name{is.witness_set}
\alias{is.witness_set}
\title{Witness_Set Object Check}
\usage{
is.witness_set(x)
}
\arguments{
\item{x}{object to be tested}
}
\value{
Vector of logicals.
}
\description{
Test whether an object is an witness_set object.
}
\examples{

if (has_bertini()) {
input <- bertini_input("x*y")
input <- modify_config(input, tracktype = 1)

# positive-dimensional run
out <- bertini(input, output = "pos_dim")

# print witness set
witness_set <- witness_set_print(out, 1, -2)
is.witness_set(witness_set)

}

}
