% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/membership_test.R
\name{membership_test}
\alias{membership_test}
\title{Membership Testing}
\usage{
membership_test(out, points, dir = tempdir(), ...)
}
\arguments{
\item{out}{the output of a positive-dimension run from bertini}

\item{points}{a matrix of points to be tested}

\item{dir}{directory to place the files in, without an ending /}

\item{...}{additional configuration arguments to be changed}
}
\value{
a 'glue' object that contains the decision of each point
}
\description{
Test if specified points lie on components of a numerical
irreducible decomposition.
}
\examples{

if (has_bertini()) {

# see ?bertini

# run initial numerical irreducible decomposition for equation xy = 0
input <- bertini_input("x*y")
input <- modify_config(input, tracktype = 1)

out <- bertini(input, output = "pos_dim")

# test if the points (0,0), (1,0), (0,1), (1,1) are in the solution set
pts_to_test <- matrix(c(0,0,1,0,0,1,1,1), nrow = 4, ncol = 2, byrow = TRUE)
membership_test(out, pts_to_test)
}
}
