% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_config.R
\name{modify-config}
\alias{modify-config}
\alias{modify_config}
\title{Add, Modify, or Remove Configuration}
\usage{
modify_config(x, ...)
}
\arguments{
\item{x}{a bertini_input object}

\item{...}{configurations to be added or modified.}
}
\value{
the bertini_input object with the added, modified, or removed
configurations
}
\description{
Add to, modify, and/or remove a configuration or multiple configurations in a
bertini_input object. This function is a thin wrapper of the
\code{\link[purrr:list_modify]{purrr::list_modify()}} function in the purrr package. Therefore, you can use
\code{\link[rlang:zap]{rlang::zap()}} to remove a configuration.
}
\examples{

(struct <- bertini_input(mp(c("x^2 + y^2 - 1","x - y"))))

modify_config(struct, precision = 124)



}
