% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_config.R
\name{modify_config}
\alias{modify_config}
\title{Add, Modify, or Remove Configuration(s) to a Bertini_Input Object}
\usage{
modify_config(x, ...)
}
\arguments{
\item{x}{a bertini_input object}

\item{...}{configurations to be added or modified.
The configurations can also be specified in a list.}
}
\value{
the bertini_input object with the added, modified,
or removed configurations
}
\description{
Add to, modify, and/or remove a configuration or multiple configurations in a
bertini_input object. This function is a thin wrapper of the
\code{list_modify} function in the purrr package. Therefore, you can
use \code{zap} to remove a configuration.
}
\examples{

struct <- bertini_input(mp(c("x^2 + y^2 - 1","x - y")))

modify_config(struct, parameterhomotopy = 1)
}
