% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_defs.R
\name{modify_defs}
\alias{modify_defs}
\title{Modify Definition(s) in a Bertini_Input Object}
\usage{
modify_defs(x, ...)
}
\arguments{
\item{x}{a bertini_input object}

\item{...}{the definitions to be modified. If you
change any part of the variable groupings, you must fully
specify the variable groupings}
}
\value{
the bertini_input object with the modified definitions
}
\description{
Modify Definition(s) in a Bertini_Input Object
}
\examples{

if (has_bertini()) {
  polys <- mp(c("x^2 + y^2 - 1", "x - y"))
  input <- bertini_input(polys)

  modify_defs(input, variable_group = "x", variable_group = "y")
}
}
