% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poly_optim.R
\name{poly_optim}
\alias{poly_optim}
\title{Polynomial Optimization}
\usage{
poly_optim(objective, constraints, varorder, ...)
}
\arguments{
\item{objective}{the objective polynomial (as a character or mpoly)}

\item{constraints}{(as a character or mpoly/mpolyList)}

\item{varorder}{variable order (see examples)}

\item{...}{stuff to pass to bertini}
}
\value{
an object of class bertini
}
\description{
Find the collection of critical points of a multivariate polynomial
unconstrained or constrained to an affine variety (algebraic set; solution
set of multivariate polynomials).
}
\examples{

if (has_bertini()) {

# unconstrained optimization of polynomial functions is available
poly_optim("x^2")
poly_optim("-x^2")
poly_optim("-(x - 2)^2")
poly_optim("-(x^2 + y^2)")
poly_optim("-(x^2 + (y - 2)^2)")
poly_optim("-(x^2 + (y - 2)^2)") # saddle

poly_optim("(x - 1) (x - 2) (x - 3)") # fix global labeling


# constrained optimization over the affine varieties is also available
# (affine variety = solution set of polynomial equations)

# find the critical points of the plane f(x,y) = x + y
# over the unit circle x^2 + y^2 = 1
poly_optim("x + y", "x^2 + y^2 = 1")

# you can specify them as a combo of mpoly, mpolyList, and characters
o <- mp("x + y")
c <- "x^2 + y^2 = 1"
poly_optim(o, c)

c <- mp("x^2 + y^2 - 1")
poly_optim(o, c)

out <- poly_optim("x + y", c)
str(out, 1, give.attr = FALSE)
cat(out$bertini_code)

# another example, note the solutions are computed over the complex numbers
poly_optim("x^2 y", "x^2 + y^2 = 3")
# solutions: (+-sqrt(2), +-1) and (0, +-sqrt(3))




}

}
