% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.bertini_input.R
\name{print.bertini_input}
\alias{print.bertini_input}
\title{Pretty Printing of Bertini_Input Objects}
\usage{
\method{print}{bertini_input}(x, silent = FALSE, ...)
}
\arguments{
\item{x}{an object of class bertini_input}

\item{silent}{logical; if TRUE, suppresses output}

\item{...}{additional parameters}
}
\value{
Invisible string of the printed object.
}
\description{
This function parses and views the bertini_input file as a valid Bertini
input file.
}
\examples{

if (has_bertini()) {

bertini_input(mp(c("x^2 + y^2 - 1", "x - y")))

}
}
