% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.poly_optim.R
\name{print.poly_optim}
\alias{print.poly_optim}
\title{Pretty printing of poly_optim (Bertini) output.}
\usage{
\method{print}{poly_optim}(x, lagrange = FALSE, digits = 3, ...)
}
\arguments{
\item{x}{an object of class poly_optim, bertini}

\item{lagrange}{show values of lagrange multipliers?}

\item{digits}{digits to round to}

\item{...}{additional parameters}
}
\value{
Invisible string of the printed object.
}
\description{
Pretty printing of poly_optim (Bertini) output.
}
\examples{

# see ?poly_optim


}
