% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.witness_set.R
\name{print.witness_set}
\alias{print.witness_set}
\title{Printing A Witness_Set Object}
\usage{
\method{print}{witness_set}(x, ...)
}
\arguments{
\item{x}{an object of class witness_set}

\item{...}{additional parameters}
}
\value{
Invisible string of the printed object.
}
\description{
Printing A Witness_Set Object
}
\examples{

# run initial numerical irreducible decomposition for
# the equation xy = 0

input <- bertini_input("x*y")
input <- modify_config(input, tracktype = 1)

out <- bertini(input, output = "pos_dim")

# print the witness set for all components in dimension 1
witness_set <- witness_set_print(out, 1, -2)
print(witness_set)
}
