% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R, R/set_paths.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{edit_r_environ}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{magrittr}{\code{\link[magrittr:pipe]{\%>\%}}}

  \item{usethis}{\code{\link[usethis:edit]{edit_r_environ}}}
}}

