% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharpening_module.R
\name{sharpening_module}
\alias{sharpening_module}
\title{Endpoint Sharpening and Postprocessing Module}
\usage{
sharpening_module(out, dir = tempdir(), quiet = TRUE, ...)
}
\arguments{
\item{out}{The output of a zero-dimensional Bertini run}

\item{dir}{directory to place the files in, without an ending /}

\item{quiet}{show bertini output}

\item{...}{additional configuration arguments to be changed.}
}
\value{
an object of class bertini
}
\description{
Endpoint Sharpening and Postprocessing Module
}
\examples{

# Examples coming soon!
}
