% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid_configurations.R
\docType{data}
\name{valid_configurations}
\alias{valid_configurations}
\title{This is a list of valid configurations}
\format{An object of class \code{character} of length 70.}
\usage{
valid_configurations
}
\description{
This is a list of valid configurations
}
\keyword{datasets}
