% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid_definitions.R
\docType{data}
\name{valid_def_names}
\alias{valid_def_names}
\title{This is a list of valid definition names}
\format{An object of class \code{character} of length 9.}
\usage{
valid_def_names
}
\description{
This is a list of valid definition names
}
\keyword{datasets}
