% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/witness_set_print.R
\name{witness_set_print}
\alias{witness_set_print}
\title{Witness Set Printing}
\usage{
witness_set_print(out, dimension, component, dir = tempdir(), ...)
}
\arguments{
\item{out}{the output of a positive-dimension run from bertini}

\item{dimension}{the dimension of the component's witness set
to be printed.}

\item{component}{The component's witness set to be printed. Indexing of
components starts at 0, not 1. To select all components of a particular
dimension, input -2.}

\item{dir}{directory to place the files in, without an ending /}

\item{...}{additional configuration arguments to be changed.}
}
\value{
an object of class witness_set
}
\description{
Print the witness set of specified component(s) of a
numerical irreducible decomposition.
}
\examples{

# run initial numerical irreducible decomposition for
# the equation xy = 0

input <- bertini_input("x*y")
input <- modify_config(input, tracktype = 1)

out <- bertini(input, output = "pos_dim")

# print the witness set for all components in dimension 1
witness_set_print(out, 1, -2)
}
