% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_homotopy.R
\name{parameter_homotopy}
\alias{parameter_homotopy}
\title{Parameter Homotopies}
\usage{
parameter_homotopy(
  input,
  parameters,
  varorder,
  precomputed_solutions = NULL,
  dir = tempdir(),
  definitions = list(),
  configurations = list(),
  subfunctions = list()
)
}
\arguments{
\item{input}{system of polynomials as either a character vector,
mpolyList, or bertini_input object}

\item{parameters}{the parameters that define the family as a character vector}

\item{varorder}{(optional) order of variables in string}

\item{precomputed_solutions}{an optional list that contains the pertinent
results of an ab initio run (start and start_parameters files) for the system. Supplying
these files can greatly decrease the computation time.}

\item{dir}{directory to place the files in, without an ending /}

\item{definitions}{an optional named list of the definitions to be given to
Bertini. The definitions name all arguments used in the polynomial and
tell Bertini what type of homotopy to use. Defaults to a total-degree homotopy.}

\item{configurations}{an optional named list of configurations to be given to
Bertini. Names will be changed to lower case.}

\item{subfunctions}{an optional named list that contains subfunctions and
other equations needed in the input file but do not need to be named in the
definitions block.}
}
\value{
a closure that accepts specific values for the parameters and computes
the solutions for those specific values
}
\description{
Parameter Homotopies
}
\examples{

# a sextic in one variable (from Bertini)
polys <- c("a0 + x*(a1 + x*(a2 + x*(a3 + x*(a4 + x*(a5 + x*a6)))))")

generator <- parameter_homotopy(polys, parameters = c("a0","a1","a2","a3","a4","a5","a6"))
generator(c(1.1, 2.4, 0.8, 3.6, -0.52, -1.8, 4.4))
}
