% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharpening_module.R
\name{sharpening_module}
\alias{sharpening_module}
\title{Endpoint Sharpening and Postprocessing Module}
\usage{
sharpening_module(out, dir = tempdir(), quiet = TRUE, ...)
}
\arguments{
\item{out}{The output of a zero-dimensional Bertini run}

\item{dir}{directory to place the files in, without an ending /}

\item{quiet}{show bertini output}

\item{...}{additional configuration arguments to be changed.}
}
\value{
an object of class bertini
}
\description{
Endpoint Sharpening and Postprocessing Module
}
\examples{

if(has_bertini()){

# initial bertini run on an example that has a high condition number
polys <- c("y - x^10", "y - 1e-10")
out <- bertini(bertini_input(polys))
out

# run the sharpening module to use the postprocessor after
# adjusting the condition number threshold (Selection 4).
# Notice that the singular solutions changed to nonsingular solutions.

sharpening_module(out, condnumthreshold = 1e12)



# root sharpening after a homotopy
polys <- c("x^2 - y^2 - 1", "x - y^2 - 1")
out <- bertini(bertini_input(polys))
print(out, digits = 20)

# sharpen roots. Select option 1.
sharpen_out <- sharpening_module(out, sharpendigits = 20)
print(out, digits = 20)
}
}
